remove improts changes route name and made methods static
