+++ packages/py-api/py_api/controllers/__init__.py
@@ -1,7 +1,7 @@
 from py_api.controllers.feature_switches_controller import FeatureSwitchesController
 from py_api.controllers.hackathon_participants_controller import ParticipantsController
+from py_api.controllers.hackathon_teams_controller import TeamsController
 from py_api.controllers.questionnaires_controller import QuestionnairesController
-from py_api.controllers.teams_controller import TeamsController
 from py_api.controllers.uploader_controller import UploaderController
 from py_api.controllers.url_shortener_controller import UrlShortenerController
 from py_api.controllers.utility_controller import UtilityController
