+++ packages/py-api/py_api/controllers/hackathon_teams_controller.py
@@ -1,6 +1,5 @@
 import json
 
-from bson.errors import InvalidId
 from bson.json_util import dumps
 from bson.objectid import ObjectId
 from fastapi.responses import JSONResponse
@@ -11,6 +10,7 @@ from py_api.utilities.parsers import filter_none_values
 
 class TeamsController:
 
+    @staticmethod
     def fetch_teams() -> JSONResponse:
         teams = list(t_col.find())
         if not teams:
@@ -28,8 +28,9 @@ class TeamsController:
                 for participant in participants
             ]
 
-        return JSONResponse(content={"teams": json.loads(dumps(teams))}, status_code=201)
+        return JSONResponse(content={"teams": json.loads(dumps(teams))}, status_code=200)
 
+    @staticmethod
     def get_team(object_id: str) -> JSONResponse:
         specified_team = t_col.find_one(filter={"_id": ObjectId(object_id)})
 
@@ -49,6 +50,7 @@ class TeamsController:
 
         return JSONResponse(content={"participant": json.loads(dumps(specified_team))}, status_code=200)
 
+    @staticmethod
     def delete_team(object_id: str) -> JSONResponse:
         delete_team = t_col.find_one_and_delete(
             filter={"_id": ObjectId(object_id)},
@@ -59,6 +61,7 @@ class TeamsController:
 
         return JSONResponse(content={"message": json.loads(dumps(delete_team))}, status_code=200)
 
+    @staticmethod
     def team_count() -> JSONResponse:
         count = t_col.count_documents({})
 
@@ -67,6 +70,7 @@ class TeamsController:
 
         return JSONResponse(content={"teams": count})
 
+    @staticmethod
     def update_team(object_id: str, update_table_model: UpdateTeam) -> JSONResponse:
         fields_to_be_updated = filter_none_values(update_table_model)
