+++ packages/py-api/py_api/models/__init__.py
@@ -3,7 +3,7 @@ from py_api.models.hackathon_participants_models import (
     NewParticipant,
     UpdateParticipant,
 )
-from py_api.models.teams_models import NewTeams, UpdateTeam
+from py_api.models.hackathon_teams_models import NewTeams, UpdateTeam
 from py_api.models.url_shortener_models import ShortenedURL
 
 __all__ = [
