+++ packages/py-api/py_api/routes/routes.py
@@ -5,9 +5,9 @@ from py_api.routes.feature_switches_routes import router as feature_switches_rou
 from py_api.routes.hackathon_participants_routes import (
     router as hackathon_participants_routes,
 )
+from py_api.routes.hackathon_teams_routes import router as teams_router
 from py_api.routes.logs_routes import router as logs_router
 from py_api.routes.questionnaires_routes import router as questionnaires_router
-from py_api.routes.teams_routes import router as teams_router
 from py_api.routes.uploader_routes import router as uploader_router
 from py_api.routes.url_shortener_routes import router as url_shortener_router
 from py_api.routes.utility_routes import router as utility_router
